+++ view/backEnd/gestionCommentaire.php
@@ -8,19 +8,22 @@
             Titre | pseudo | contenu | Date d'ajout
         </li>
         <?php
-        foreach($commentaires as $commentaire)
-        {
-            echo'<li>'.$commentaire->getNomChapitre().' | '.$commentaire->getPseudo().'
-			| '.$commentaire->getContenu().' | '.$commentaire->getDateAjout().'
-			<a href="index.php?pages=backEnd&page=gestionCommentaires&action=supprimerCommentaire&commentaire='.$commentaire->getId().'">supprimer commentaire</a>
+        if (isset($commentaires)) {
+            foreach ($commentaires as $commentaire) {
+                echo '<li>' . $commentaire->getNomChapitre() . ' | ' . $commentaire->getPseudo() . '
+			| ' . $commentaire->getContenu() . ' | ' . $commentaire->getDateAjout() . '
+			<a href="index.php?pages=backEnd&page=gestionCommentaires&action=supprimerCommentaire&commentaire=' . $commentaire->getId() . '">supprimer commentaire</a>
 			 | ';
-            if ($commentaire->getSignalement()) {
-                echo '<a href="index.php?pages=backEnd&page=gestionCommentaires&action=desactiverSignalement&commentaire='.$commentaire->getId().'">supprimer signalement</a></li>';
-            }else {
+                if ($commentaire->getSignalement()) {
+                    echo '<a href="index.php?pages=backEnd&page=gestionCommentaires&action=desactiverSignalement&commentaire=' . $commentaire->getId() . '">supprimer signalement</a></li>';
+                } else {
 
-            echo 'aucun signalement';
+                    echo 'aucun signalement';
+                }
+
             }
-
+        }else{
+            echo'aucun commentaire';
         }
         ?>
     </ul>
